refactor : 공통 레이아웃 mypage 적용
